requirements/test.txt
-r prod.txt
black>=19.10b0
coverage>=6.1.2
faker
flake8>=4.0.1
ghp-import
linkify-it-py
myst-parser
pre-commit>=2.15.0
pytest>=6.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.1
Sphinx>=4.3.0
sphinx-autodoc-typehints
sphinx-click
twine>=3.6.0
wheel>=0.37.0
